Newest Mozilla Clubs network certificates
This is an option for Mozilla Clubs Captain remix and issue certificate for club members, the design is based on Mozilla Network website (https://network.mozilla.org/). Please touch the Mozilla Clubs staff team to get instructions on the text, and don't issue this as Mozilla Corporation ou Foundation, issue this as local Mozilla Club based on your city and inspired by Mozilla Leadership Network.